requirements.txt
contourpy==1.3.0
cycler==0.12.1
fonttools==4.57.0
importlib_resources==6.5.2
kiwisolver==1.4.7
matplotlib==3.9.4
numpy==2.0.2
packaging==24.2
pillow==11.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
scipy==1.13.1
six==1.17.0
zipp==3.21.0
h5py
PyYAML
pyvista
